+Navigation +UserScreen

+++ app/src/main/java/com/example/users/data/source/local/room/UserEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "Users")
 data class UserEntity(
     @PrimaryKey(autoGenerate = true)
-    val roomId: Long? = null,
+    val roomId: Int? = null,
 
     val id: String,
     val avatarUrl: String,
